Response stores score not FK on available_response
I'm changing the available responses for the questionnaires because I'm finding them difficult to answer consistently. I'm therefore updating their labels. In order to maintain the previous data, I'm modifying the way responses get stored. Instead of recording the ID of the response (as a FK), I'm going to simply store the score of the response. This is done as a simple `IntegerField()` instead of a FK.